refactor job master process
